requirements.txt
argostranslate
faiss-cpu
ir_datasets
numpy
pandas
spacy